grades fixed

+++ LMS_handout/LMS/Controllers/ProfessorController.cs
@@ -504,7 +504,7 @@ namespace LMS.Controllers
             string semester = season + year.ToString();
             string class_grade;
             float? grade = 0;
-            UInt32? weight_sum = 0;
+            float? weight_sum = 0;
             var all_categories =
                 from c in db.AssignmentCategories
                 where c.ClassNavigation.Course.Number == num
@@ -542,8 +542,8 @@ namespace LMS.Controllers
                 if (values_sum > 0)
                     grade += (((float)score_sum / (float)values_sum) * current_category.Weight);
             }
-            grade = grade * (100 / weight_sum);
-
+            grade = grade * 100f / (float)weight_sum;
+            
             if (grade >= 93)
                 class_grade = "A";
             else if (grade >= 90 && grade < 93)
